requirements.txt
absl-py==2.2.2
accelerate==0.34.0
aiofiles==23.2.1
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiohttp-cors==0.7.0
aiosignal==1.3.1
airportsdata==20241001
annotated-types==0.7.0
anthropic==0.48.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
astor==0.8.1
asttokens==2.4.1
attrs==24.1.0
auto-gptq==0.6.0
av==14.0.1
backcall==0.2.0
beautifulsoup4==4.13.4
blake3==1.0.4
bleach==6.2.0
blobfile==3.0.0
cachetools==5.5.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.31.2
coloredlogs==15.0.1
colorful==0.5.6
comm==0.2.2
compressed-tensors==0.8.1
contourpy==1.2.1
cramjam==2.9.0
cycler==0.12.1
datasets==2.20.0
debugpy==1.8.5
decorator==5.1.1
decord==0.6.0
deepspeed==0.14.4
defusedxml==0.7.1
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
docopt==0.6.2
docstring_parser==0.16
einops==0.8.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.112.0
fastjsonschema==2.21.1
fastparquet==2024.11.0
ffmpy==0.4.0
filelock==3.17.0
fire==0.6.0
flash-attn==2.6.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
gekko==1.2.1
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-auth==2.38.0
googleapis-common-protos==1.66.0
gradio==4.40.0
gradio_client==1.2.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.7
imageio==2.37.0
importlib_metadata==8.4.0
importlib_resources==6.4.0
iniconfig==2.0.0
interegular==0.3.3
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
langchain==0.2.7
langchain-core==0.2.12
langchain-openai==0.1.14
langchain-text-splitters==0.2.2
langgraph==0.1.5
langsmith==0.1.147
lark==1.2.2
linkify-it-py==2.0.3
-e git+https://github.com/FreedomIntelligence/LLaMA-Factory@752373eb11ea9af04cd137d4a2514c803b7477ba#egg=llamafactory
llvmlite==0.43.0
lm-format-enforcer==0.10.1
loguru==0.7.2
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.15.0
mistral_common==1.5.2
mistune==3.1.3
mpmath==1.3.0
msgpack==1.0.8
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
mysql-connector-python==9.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.17.0
onnxsim==0.4.36
openai==1.60.1
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
optimum==1.23.3
orjson==3.10.6
outlines==0.0.34
outlines_core==0.1.26
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post4
peft==0.12.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.2
pluggy==1.5.0
portalocker==3.1.1
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyairports==2.1.1
pyarrow==18.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycountry==24.6.1
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.1.2
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.2.0
ray==2.35.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rouge==1.0.1
rouge-chinese==1.0.3
rpds-py==0.20.0
rsa==4.9
ruff==0.5.6
safetensors==0.4.3
scandir==1.10.0
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.12.0
setproctitle==1.3.3
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sse-starlette==2.1.3
stack-data==0.6.2
starlette==0.37.2
sty==1.0.6
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
text2vec==1.3.0
textual==1.0.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tiktoken==0.7.0
timeout-decorator==0.5.0
tinycss2==1.4.0
tokenizers==0.20.3
tomlkit==0.12.0
torch==2.3.0
torchsummary==1.5.1
torchvision==0.18.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.2
triton==2.3.0
trl==0.9.6
typer==0.12.3
typing_extensions==4.12.2
tyro==0.8.5
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.20.0
validators==0.34.0
virtualenv==20.29.1
vllm==0.4.3
vllm-flash-attn==2.5.8.post2
-e git+ssh://git@github.com/open-compass/VLMEvalKit.git@aaa1420ed21f0f4a9fb6d355090b99799496de16#egg=vlmeval
wandb==0.17.5
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.1.3
wrapt==1.17.2
xformers==0.0.26.post1
xgrammar==0.1.11
XlsxWriter==3.2.2
xxhash==3.4.1
yarg==0.1.9
yarl==1.9.4
zipp==3.20.1
zstandard==0.23.0
